[YTDB-558] Move createLegacySqlRid to MatchLiteralBuilder
The RecordIdInternal -> SQLRid legacy converter lived in StartStepRecogniser (the Gremlin-to-MATCH start-step recognizer) but was consumed by MatchLiteralBuilder in the SQL/MATCH executor via a static import, making the SQL layer depend on the Gremlin translator for a conversion unrelated to Gremlin. Move it to MatchLiteralBuilder, whose contract already covers RecordIdInternal -> SQLRid literal building, so the dependency now runs the correct direction: the translator depends on the SQL builder it targets. StartStepRecogniser calls MatchLiteralBuilder.createLegacySqlRid directly; the redundant private toSqlRid delegate and the now-unused SQLInteger import are gone. Behavior unchanged.
